Apply swiftformat
I had issues with indenting mainly, so I introduced `swiftformat`.
This way, we will have a shared style.
I tried to match the current style as much as possible.
There is one litte problem I can't work around, it's the fact that `swiftformat` indents the first and last line of a comment block (in some cases).