Refactor sidebar (#60)
* refactor: Move content screen states from Sidebar into MainScreen

* chore: Rename AvatarRow back to ContactRow

* style: Run swiftformat after merging

* chore: Move SpotlightGroupBoxStyle into ProseUI to make it available for other targets

* chore: Grey out groups section in sidebar

* chore: Rename State/Action typealiases

* chore: Replace CasePath operator with initializer

* fix: Prevent screen flash when changing selection in sidebar

When pressing a row in the Sidebar and dragging the mouse without releasing the mouse button the SwiftUI List updated the selection multiple times without actually changing it. So the binding was written to multiple times with the same value. Now instead of reacting to the action that is triggered when writing to the binding, we’ll observe the selection property and only change screens when after it changed.